requirements-dev.txt
astroid==2.2.5
autopep8==1.4.4
ddt==1.2.1
gitdb==0.6.4
gitdb2==2.0.5
GitPython==2.1.12
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.5.0
pylint==2.3.1
six==1.12.0
smmap==0.9.0
smmap2==2.0.5
typed-ast==1.4.0
wrapt==1.11.2
